[GraphMaker] Fix external edges
