+++ GraphMaker/app/data/PlanData.py
@@ -1,3 +1,81 @@
+from app.general.constants import *
+
+class Node:
+    def __init__(self, nb, coords, access_points, aliases=tuple()):
+        self.nb = nb
+        self.coords = coords
+        self.access_points_ = access_points
+        self.color = 'green' if access_points is not None else 'red'
+        self.aliases_ = list(aliases)
+
+    def coord(self, c=None):
+        if c is None:
+            return self.coords
+        else:
+            self.coords = c
+
+    def id(self):
+        return self.nb
+
+    def access_points(self, ap=None):
+        if ap is None:
+            return self.access_points_
+        else:
+            self.access_points_ = ap
+
+    def aliases(self, a=None):
+        if a is None:
+            return self.aliases_
+        else:
+            self.aliases_ = list(a)
+
+    def text(self, nb_tab=0):
+        text = (TAB * (nb_tab+1)) + '<coord x="{}" y="{}" />\n'.format(*self.coord())
+        if self.access_points() is not None:
+            text += self.access_points().text(nb_tab+1)
+        if self.aliases() and len(self.aliases()) > 0:
+            text += (TAB*(nb_tab+1)) + '<aliases>\n'
+            for alias in self.aliases():
+                text += (TAB*(nb_tab+2)) + '<alias>{}</alias>\n'.format(alias)
+            text += (TAB*(nb_tab+1)) + '</aliases>\n'
+        return '{0}<node id="{1}">\n{2}{0}</node>\n'.format(TAB*nb_tab, self.id(), text)
+
+class Edge:
+    def __init__(self, weight, coords, extremity_ids):
+        self.weight_ = weight
+        self.coords = coords
+        self.extremity_ids = extremity_ids
+
+    def coord(self, c=None):
+        if c is None:
+            return self.coords
+        else:
+            self.coords = c
+
+    def weight(self, w=None):
+        if w is None:
+            return self.weight_
+        else:
+            self.weight_ = w
+
+    def text(self, nb_tab=0):
+        return (TAB*nb_tab) + '<edge beg="{}" end="{}" weight="{}" />\n' \
+                              .format(*self.extremity_ids, self.weight())
+
+class ExternalEdge(Edge):
+    def __init__(self, weight, extremity_ids, plan):
+        super().__init__(weight, [0, 0], extremity_ids)
+        self.plan = plan
+
+    def extremities(self, ext=None):
+        if ext is not None:
+            self.extremity_ids = ext[:]
+        else:
+            return self.extremity_ids
+
+    def text(self, nb_tab=0):
+        return (TAB*(nb_tab)) + '<edge beg="{}" end="{}" weight="{}" plan="{}">' \
+                                .format(*self.extremity_ids, self.weight(), self.plan)
 
 class PlanData:
     def __init__(self):
